just save this

+++ rl_modules/replay_buffer.py
@@ -22,7 +22,7 @@ class replay_buffer:
         self.sym_image = sym_image
         # create the buffer to store info
         self.buffers = {'obs_states': np.empty([self.size, self.T + 1, self.env_params['obs']]),
-                        'obs_imgs': np.empty([self.size, self.T + 1, 100, 100, chan_size], dtype=np.uint8),
+                        'obs_imgs': np.empty([self.size, self.T + 1, 100, 100, chan_size], dtype=np.float32),
                         'ach_goal_states': np.empty([self.size, self.T + 1, self.env_params['goal']]),
                         'goal_states': np.empty([self.size, self.T+1, self.env_params['goal']]),
                         'actions': np.empty([self.size, self.T, self.env_params['action']]),
@@ -149,7 +149,7 @@ class new_replay_buffer:
         env_state_dim = 5
         # create the buffer to store info
         self.buffers = {'obs_states': np.empty([self.size, self.T + 1, self.env_params['obs']]),
-                        'obs_imgs': np.empty([self.size, self.T + 1, height, width, chan_size], dtype=np.uint8),
+                        'obs_imgs': np.empty([self.size, self.T + 1, height, width, chan_size], dtype=np.float32),
                         'ach_goal_states': np.empty([self.size, self.T + 1, self.env_params['goal']]),
                         'goal_states': np.empty([self.size, self.T+1, self.env_params['goal']]),
                         'actions': np.empty([self.size, self.T, self.env_params['action']]),
